backend/requirements.txt
flask
torch
torchaudio
opencv-python
ffmpeg-python
numpy
pydub
transformers
mediapipe
